Refactor: tidy up bash scripts

+++ macOS/brews.sh
@@ -1,47 +1,13 @@
 #!/bin/bash
 
-# ---------------------------------------------------
-# Utility functions
-# ---------------------------------------------------
-# Pretty print a message.
-function log {
-  echo -e "\n"
-  echo "============================================="
-  echo $1
-  echo "============================================="
-}
+DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" > /dev/null 2>&1 && pwd )"
 
-# Check if a cmd exists.
-function cmd_exists {
-  command -v $1 > /dev/null 2>&1
-}
+source $DIR/../config/variables.sh
+source $DIR/../lib/utils.sh
 
-# Log a skipping message.
-function skip {
- log "$1 already exists. Skipping"
-}
-
-# Install brew formula
-function install_formula {
-  if cmd_exists $1; then
-    skip $1
-  else
-    log "Installing $1"
-    brew install $1
-  fi
-}
-
-
-# ---------------------------------------------------
-# Variables
-# ---------------------------------------------------
-EMAIL="sajadtorkamani1@gmail.com"
-DOTFILES_DIR=$HOME/.config/dotfiles
-DEFAULT_RUBY_VERSION="2.6.3" # The Ruby version to install
-DEFAULT_NODE_VERSION="v10.16.0"
-
-
-# Install Homebrew
+# --------------------------------------------------
+# Homebrew
+# --------------------------------------------------
 if cmd_exists "brew"; then
   skip "Homebrew"
 else
@@ -49,13 +15,16 @@ else
   /usr/bin/ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
 fi
 
-if [ -e $HOME/.ssh/id_rsa ]; then
+if [[ -e $HOME/.ssh/id_rsa ]]; then
   skip "SSH keys"
 else
   log "Generating SSH key"
   ssh-keygen -t rsa -b 4096 -C $EMAIL
 fi
 
+# --------------------------------------------------
+# ZSH
+# --------------------------------------------------
 if [[ $SHELL =~ "zsh" ]]; then
   skip "ZSH"
 else
@@ -65,6 +34,9 @@ else
   chsh -s $(which zsh)
 fi
 
+# --------------------------------------------------
+# rbenv
+# --------------------------------------------------
 if cmd_exists "rbenv"; then
   skip "rbenv"
 else
@@ -73,16 +45,22 @@ else
   gem install bundler
 fi
 
+# --------------------------------------------------
+# nvm
+# --------------------------------------------------
 if [[ -e $HOME/.nvm ]]; then
   skip "nvm"
 else
   log "Installing nvm"
   curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.34.0/install.sh | bash
   export NVM_DIR="$HOME/.nvm"
-  [ -s "$NVM_DIR/nvm.sh" ] && \. "$NVM_DIR/nvm.sh"  # This loads nvm
+  [[ -s "$NVM_DIR/nvm.sh" ]] && \. "$NVM_DIR/nvm.sh"  # This loads nvm
   nvm install $DEFAULT_NODE_VERSION
 fi
 
+# --------------------------------------------------
+# Yarn
+# --------------------------------------------------
 if cmd_exists "yarn"; then
   skip "yarn"
 else
@@ -90,9 +68,18 @@ else
   brew install yarn --ignore-dependencies
 fi
 
+# --------------------------------------------------
+# AWS CLI
+# --------------------------------------------------
+if cmd_exists "aws"; then
+  skip "aws"
+else
+  pip3 install awscli --upgrade --user
+fi
 
-pip3 install awscli --upgrade --user
-
+# --------------------------------------------------
+# Brews
+# --------------------------------------------------
 install_formula composer
 install_formula ffmpeg
 install_formula httpd
